Added initial Tour UI, which could still use some work.  Also cleaned up code in several places.
The Tour UI still needs a landscape layout (coming in a later commit).  We also need to insert a MapFragment with all of the places on the tour marked in the large empty space above the tour details.

PlaceListAdapter renamed to PlaceCursorAdapter to reflect that it now uses Cursors instead of Lists.